Change type of view_t::window_cells size_t => int
It was cast to int almost on every use.

+++ src/flist_hist.c
@@ -221,7 +221,7 @@ flist_hist_lookup(view_t *view, const view_t *source)
 	view->list_pos = pos;
 	if(rel_pos >= 0)
 	{
-		view->top_line = pos - MIN((int)view->window_cells - 1, rel_pos);
+		view->top_line = pos - MIN(view->window_cells - 1, rel_pos);
 		if(view->top_line < 0)
 			view->top_line = 0;
 		view->curr_line = pos - view->top_line;
@@ -229,7 +229,7 @@ flist_hist_lookup(view_t *view, const view_t *source)
 	else
 	{
 		const int last = fpos_get_last_visible_cell(view);
-		if(view->list_pos < (int)view->window_cells)
+		if(view->list_pos < view->window_cells)
 		{
 			scroll_up(view, view->top_line);
 		}
